+++ elegy/generalized_module/linen_module_test.py
@@ -84,7 +84,8 @@ class TestLinenModule(unittest.TestCase):
                 return x
 
         model = elegy.Model(ModuleA())
-
+        model.init(x=jnp.ones([10, 2]))
+        
         summary_text = model.summary(x=jnp.ones([10, 2]), depth=1, return_repr=True)
         assert summary_text is not None
