+++ elegy/nn/linear_test.py
@@ -21,5 +21,5 @@ class LinearTest(TestCase):
 
         x = np.ones([3, 5])
 
-        y_pred = model.predict(x)
+        y_pred = model.predict(x, init=True)
         logs = model.evaluate(x)
